requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
black==21.7b0
click==8.0.1
Django==3.2.6
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
